+++ core/src/main/scala/scalafix/Scalafix.scala
@@ -5,7 +5,7 @@ import scala.meta._
 import scala.meta.inputs.Input
 import scala.meta.parsers.Parsed
 import scalafix.rewrite.RewriteCtx
-import scalafix.rewrite.SemanticApi
+import scalafix.rewrite.ScalafixMirror
 import scalafix.util.AssociatedComments
 import scalafix.util.Patch
 import scalafix.util.TokenList
@@ -13,7 +13,7 @@ import scalafix.util.TokenList
 object Scalafix {
   def fix(code: Input,
           config: ScalafixConfig,
-          semanticApi: Option[SemanticApi]): Fixed = {
+          semanticApi: Option[ScalafixMirror]): Fixed = {
     config.parser.apply(code, config.dialect) match {
       case Parsed.Success(ast) =>
         fix(ast, config, semanticApi)
@@ -23,7 +23,7 @@ object Scalafix {
   }
   def fix(ast: Tree,
           config: ScalafixConfig,
-          semanticApi: Option[SemanticApi]): Fixed = {
+          semanticApi: Option[ScalafixMirror]): Fixed = {
     val tokens = ast.tokens
     implicit val ctx = RewriteCtx.fromCode(ast, config, semanticApi)
     val patches = config.rewrites.flatMap(_.rewrite(ast, ctx))
